undo changes (mainly CLI) irrelevant to API PR
