+++ cli/integration/tests/waiter/util.py
@@ -403,16 +403,3 @@ def get_instances_not_in_output(possible_instances, stdout_output):
     return [instance
             for instance in possible_instances
             if instance['id'] not in stdout_output]
-
-def get_specific_instances_for_service(waiter_url, service_id, instance_type, expected_status_code=200):
-    """returns specific instances map of a service"""
-    headers = {
-        'Content-Type': 'application/json',
-        'x-cid': cid()
-    }
-    response = session.get(f'{waiter_url}/apps/{service_id}', headers=headers)
-    service = response.json()
-    assert expected_status_code == response.status_code, \
-        f'Expected {expected_status_code}, got {response.status_code} with body {response.text}'
-    return service['instances'][instance_type]
-
